# Do not use buffer pool for ParseSIP