@emiago 2024-10-19
@dennwc I agree. I think I had this when there were concurent parsing some old code. So copy of bytes was needed. Now all is more synced.